chore: Rewrote tests with mocha
