+++ test-node/tests/lib/broccoli/translation-reducer/linter/index/base-case-1-test.js
@@ -1,0 +1,22 @@
+import { expect } from 'chai';
+import Linter from 'ember-intl/lib/broccoli/translation-reducer/linter/index.js';
+
+describe('lib | broccoli | translation-reducer | linter | index', function () {
+  it('base case (1)', function () {
+    const linter = new Linter({
+      requiresTranslation() {
+        return true;
+      },
+    });
+
+    const output = linter.lint({
+      de: {},
+      en: {},
+    });
+
+    expect(output).to.deep.equal({
+      icuMismatch: [],
+      missingTranslations: [],
+    });
+  });
+});
