--- test-node/tests/lib/broccoli/translation-reducer/linter/index/we-cannot-match-icu-arguments-by-format-options.test.ts
@@ -1,28 +1,0 @@
-// @ts-expect-error: Could not find a declaration file
-import Linter from 'ember-intl/lib/broccoli/translation-reducer/linter/index.js';
-
-import { assert, test } from '../../../../../helpers/index.js';
-
-test('lib | broccoli | translation-reducer | linter | index > we cannot match ICU arguments by format options', function () {
-  const linter = new Linter({
-    requiresTranslation() {
-      return true;
-    },
-  });
-
-  const output = linter.lint({
-    de: {
-      'key-with-date': 'Es ist jetzt {timestamp, date, long}.',
-      'key-with-time': 'Es ist jetzt {timestamp, time, long}.',
-    },
-    en: {
-      'key-with-date': 'It is now {timestamp, time, long}.',
-      'key-with-time': 'Es ist jetzt {timestamp, time, short}.',
-    },
-  });
-
-  assert.deepStrictEqual(output, {
-    icuMismatch: [],
-    missingTranslations: [],
-  });
-});
